software/pyproject.toml
[project]
name = "lorem-eth"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Matias",email = "matias@ufscar.br"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-lorem (>=1.3.0.post3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
